requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
PyMySQL==1.1.0
Werkzeug==2.3.7
Jinja2==3.1.2
python-dotenv==1.0.0
reportlab==4.0.7
